Get rid of setjmp/longjmp hack, yay!
As expected, it didn't take long for setjmp/longjmp pair to cause us
problems. After a trivial change in 9694c851c76b9764 the app crashes
again on exit due to longjmp() crossing thread-boundary.

This changes the hack with a call back to Java side, which then
initiates exit of Activity.