Automatically generate signatures for JNI functions
Before building native code, call javap to get the signatures of the
functions declared as native in Java code. This allows us having the
simplest possible declaration of native functions.

Unfortunately, that involves a little bit of awk magic. The first call
to awk filters out the pairs of lines for each native function
declaration. The second one transforms those lines into a C variable
declaration, of the form

    char const* jni_signature_functionName =
        "<the JNI signarure>";