Make message handlers static inner classes
Lint suggested this could avoid leaked objects in certain cases.

If the handler class is not static, it will have a reference to the
outer object (usually Activity), and the messages will have a reference
to a target handler. So in a rare case of message getting stalled in the
queue for a long time, GC will be unable to collect the heavy-weight
objects.

Explanation and solutions here:
http://stackoverflow.com/questions/11407943/this-handler-class-should-be-static-or-leaks-might-occur-incominghandler
http://groups.google.com/group/android-developers/msg/959f98f16a038694?pli=1