allow commands with spaces in the quickbar
The quickbar currently breaks quickbar commands into a command and a caption at the first space. This makes it break it on the last space so that you can use commands with space in them (e.g. :new /mnt/sdcard/long/path/to/my/file) as long as you include a caption